fix: Adapter missing depending on require order
After fixing the defined? checks, all pagination adapters need to be
required before this gem for it to work at all, which is cumbersome.
Instead, we remove the defined? check completely and rely only on the
duck-test already present.

This has worked fine all for all the past as far as I know.